wales: apply suggestions from review
